backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "curl-cffi>=0.14.0",
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-rest-passwordreset>=1.5.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "lxml>=6.0.2",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.1",
    "selenium>=4.41.0",
    "selenium-stealth>=1.0.6",
    "webdriver-manager>=4.0.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
